pyproject.toml
[project]
name = "pkcs11_learning"
version = "0.0.1"
authors = [
    { name = "Andrea Stedile", email = "andrea.stedile@unitn.it" },
]
description = "PKCS #11 model learning"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    # "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-pkcs11",
    "aalpy",
    "pycryptodome"
]

[project.urls]
Homepage = "https://github.com/andreastedile/pkcs11-learning"
Issues = "https://github.com/andreastedile/pkcs11-learning/issues"
